Deprecate Configuration filtered File and FileCollection methods
As part of idiomatic JVM, we want to simplify Gradle's API to reduce the number of ways to accomplish a single operation.
ArtifactView can perform the same behaviors of this API, and more, and is actively being maintained and updated with new features.
As such, we are deprecating this API for removal in 9.0