Allow variant reselection to reselect variants referenced by dependencies with artifacts
Previously, if a dependency with an explicit artifact referenced a variant, upon reselection, the explicit artifact was never reselected.
This change resolves this by configuring the reselection variants in much the same way as normal. However, instead of including all sub-variants
of the curerntly selected variant (with the artifact dependecy referncing it), we only include the adhoc variant for the user-requested artifact.

This means that if the attributes do not change, the requested artifact will be selected again. Otherwise, we can continue to select the other variants
in the component.

As part of this change, we make the following notable other changes
- Split DefaultArtifactSet into ResolvedVariantArtifactSet and DefaultResolvedVariantSet
- Allow 'selectFromAllVariants' to be specified at the time of selection as opposed to upon construction of the variant selector
- Remove the concept of a 'legacy' variant set from ResovledVariantSet and instead migrate that functionality to ResolvedVariantArtifactSet
- Significant updates to ResovledArtifactsGraphVisitor and ComponentGraphResolveState implementations to ensure variant state objects don't have to 'know' about the variants of its owning component
- Rename ArtifactSelector to VeriantArtifactResovler