Improve CC hit rate for buildCache/integration-tests snippet
Prior to this commit, the build script read AGENT_NUMBER environment
variable at configuration time, making it a configuration input.
With providers, we only read the variable at execution time.

It is even easier to call `System.getenv` in the `asArguments` method
but we miss opportunity to describe `@Internal` use this way.